develop fix. trade modify

+++ Back/selly-trade-service/src/main/java/com/nct/sellytradeservice/model/repository/TradeLogRepository.java
@@ -30,7 +30,7 @@ public interface TradeLogRepository extends JpaRepository<TradeLog, Long> {
   HistoryDto selectDateHistory(@Param("historyDate") String date, @Param("articleId") Long articleId);
 
   @Query(value = "SELECT AVG(tradePrice) AS avgPrice FROM TradeLog WHERE articleId = :articleId", nativeQuery = true)
-  Double selectTotalAvg(@Param("articleId") Long articleId);
+  Float selectTotalAvg(@Param("articleId") Long articleId);
 
 
   TradeLog findTopByArticleIdAndTradeTimeBetweenOrderByTradeTimeDesc(Long articleId, LocalDateTime startDateTime, LocalDateTime endDateTime);
